requirements.txt
# GESTIMA 1.0 - Python dependencies

# === WEB FRAMEWORK ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# === DATABÁZE ===
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
greenlet>=3.0.0
alembic>=1.13.0

# === VALIDACE ===
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# === EXCEL ===
pandas>=2.2.0
openpyxl>=3.1.2

# === SECURITY ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
slowapi>=0.1.9

# === AI SERVICES ===
anthropic>=0.40.0  # Claude AI for quote request parsing
pymupdf>=1.24.0  # PDF parsing and manipulation
pypdf>=4.0.0  # PDF reading and merging
reportlab>=4.0.0  # PDF generation for annotations

# === UTILS ===
python-dotenv>=1.0.0
httpx>=0.26.0

# === DEV & TESTS ===
pytest>=8.0.0
pytest-asyncio>=0.23.0
